refactor: Move db fixtures into its own module
